+++ CHANGELOG.md
@@ -18,6 +18,8 @@ SPDX-License-Identifier: MIT
   - Use one snapshot per test case and rustdoc version.
 - Refactor member lookup for compatibility with Rust 1.46.0.
 - Add tests for Rust 1.46.0.
+- Improve test suite:
+  - Add test for `Parser::find_member`.
 
 ## v0.2.0 (2020-08-11)